+++ res/CHANGELOG.md
@@ -17,6 +17,7 @@
 * Use all 3 measured currents in sampled data instead of calculating one.
 * Added support for loading sampled data CSV files.
 * Always remove trailing whitespaces on save in CodeEditor.
+* Comment block fix on windows and general improvements.
 
 ---